requirements.txt
# Core Dependencies
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# AI Models & Processing
faster-whisper>=0.10.0
ffmpeg-python>=0.2.0
gTTS>=2.5.0  # For test audio generation

# API Communication
httpx>=0.26.0

# Security & Configuration
keyring>=24.3.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools
black>=24.1.0
isort>=5.13.0
pylint>=3.0.0
mypy>=1.8.0

# Note: GUI framework (PyQt5/PySide6) will be installed separately when building GUI
